Reduced amount of variables & variable sizes in structs, also marked explicit amounts for each array. Same error

+++ contracts/minter.sol
@@ -201,7 +201,7 @@ contract Minter is ERC721Enumerable, IMinter, VRFConsumerBase {
         for(uint8 i =0; i< _amount; i++){
             totalMintSupply += 1;
             _mint(_msgSender(), totalMintSupply);
-            raptorStats[totalMintSupply] = Stats(1,1,1,0,0,0,0,0,0,0,0,0,0,0,block.timestamp,false);
+            raptorStats[totalMintSupply] = Stats(1,1,1,0,0,0,0,0,0,0,0,0,0,0,uint32(block.timestamp),false);
             _approve(gameAddress, totalMintSupply);
             emit Mint(_msgSender(), totalMintSupply);
             if(totalMintSupply == totalLimit) {
@@ -237,7 +237,7 @@ contract Minter is ERC721Enumerable, IMinter, VRFConsumerBase {
                 totalMintSupply +=1;
                 _mint(rewardedAddresses[i], totalMintSupply);
                 _approve(gameAddress, totalMintSupply);
-                raptorStats[totalMintSupply] = Stats(1,1,1,0,0,0,0,0,0,0,0,0,0,0,block.timestamp,true);
+                raptorStats[totalMintSupply] = Stats(1,1,1,0,0,0,0,0,0,0,0,0,0,0,uint32(block.timestamp),true);
             }
         }
     }
